--- src/components/molecules/FormFields/FilesInput/Input.module.css
@@ -1,3 +1,0 @@
-.input {
-  composes: input from '../../../atoms/Input/InputElement.module.css';
-}
